requirements.txt
audioread==3.0.0
chardet==5.1.0
charset-normalizer==3.1.0
customtkinter==5.1.3
fastapi==0.99.0
ffmpeg-python==0.2.0
Flask==2.1.3
Flask-Cors==3.0.10
Flask-RESTful==0.3.10
Flask-SocketIO==4.3.2
flask-talisman==1.0.0
h5py==3.9.0
joblib==1.3.1
nltk==3.8.1
numba==0.57.1
numpy==1.24.4
openai==0.27.6
openai-whisper==20230314
openpyxl==3.1.2
pandas==1.5.3
pip==23.1.2
protobuf==3.20.0
pyarrow==12.0.1
PyAudio==0.2.13
PyAudioWPatch==0.2.12.5
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.10.9
pypinyin==0.44.0
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
pyqt5-tools==5.15.9.3.3
pyreadline3==3.4.1
pytorch-wpe==0.0.1
pytz==2023.3
PyYAML==6.0
QDarkStyle==3.1
qt-material==2.14
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
QtPy==2.3.1
regex==2023.6.3
requests==2.31.0
responses==0.18.0
scikit-learn==1.3.0
scipy==1.11.1
sentencepiece==0.1.99
setuptools==67.8.0
setuptools-rust==1.6.0
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.3.5
starlette==0.27.0
sympy==1.12
tensorboard==1.15.0
text-unidecode==1.3
TextGrid==1.5
threadpoolctl==3.1.0
tiktoken==0.3.1
tls-client==0.2.1
tokenizers==0.13.2
tomli==2.0.1
torch==2.0.1
torch-complex==0.4.3
torchaudio==2.0.2
tqdm==4.65.0
transformers==4.26.1
urllib3==2.0.3
Wave==0.0.2
websocket-client==1.6.1
